+++ src/main/java/delete_flashcard_use_case/FcRInputBoundary.java
@@ -1,4 +1,4 @@
-package remove_flashcard_use_case;
+package delete_flashcard_use_case;
 /**
  * Input boundary for flashcard remover.
  * Application business rules.
