+++ src/main/java/entities/Flashcard.java
@@ -5,8 +5,8 @@ import java.time.*;
 public class Flashcard {
     private String term;
     private String definition;
-    private LocalDateTime creationDate;
-    private int flashcardId;
+    private final LocalDateTime creationDate;
+    private final int flashcardId;
     private int belongsToId;
 
 
@@ -33,14 +33,6 @@ public class Flashcard {
 
     public int getFlashcardId() {
         return flashcardId;
-    }
-
-    public int getBelongsToId() {
-        return belongsToId;
-    }
-
-    public void setBelongsToId(int belongsToId) {
-        this.belongsToId = belongsToId;
     }
 
     public void setDefinition(String definition) {
