+++ src/main/java/entities/FlashcardSetFactory.java
@@ -1,7 +1,5 @@
 package entities;
 
-import java.util.List;
-
 public class FlashcardSetFactory {
     public FlashcardSet create(String title, String description, boolean isPrivate,
                         int flashcardSetId, String ownerUsername) {
