+++ src/main/java/login_and_signup_use_case/login_and_signup_use_case_screens/LabelTextPanel.java
@@ -3,7 +3,7 @@ package login_and_signup_use_case.login_and_signup_use_case_screens;
 import javax.swing.*;
 
 /**
- * Creating the ability to have a text panel to take have that label and textfield
+ * Creating the ability to have a text panel to take have that label and text-field
  *<p>
  * Frameworks & Drivers Layer
  * @author Aryan Chablani (with inspiration from Professor Paul Gries)
