+++ src/main/java/login_and_signup_use_case/login_and_signup_use_case_screens/RegisterScreen.java
@@ -10,7 +10,7 @@ import java.awt.event.ActionListener;
 import java.io.IOException;
 
 /**
- * Create a swing user interface for the Sign Up Screen
+ * Create a swing user interface for the Sign-Up Screen
  *<p>
  * Frameworks and Drivers
  * @author Aryan Chablani (with inspiration from Professor Paul Gries)
