+++ src/main/java/search_use_case/SearchController.java
@@ -1,9 +1,6 @@
 package search_use_case;
 
-import login_and_signup_use_case.UserLoginResponseModel;
-
 import java.io.IOException;
-import java.util.ArrayList;
 /**
  * Search Controller
  * <p>
